% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCutoff.R
\name{lcv}
\alias{lcv}
\title{Lico Critical Value}
\usage{
lcv(n)
}
\arguments{
\item{n}{sample size (integer)}
}
\value{
critical value
}
\description{
`lcv` computes the critical value for computing cut-off values.
This is based on Smith et al. (1998) that suggested refined cuttoff values
tthat take into account weighted discrepancy. It is given by:

\eqn{critical.value = 1 + \frac{2}{\sqrt{n}}}
}
\examples{
lcv(200)
}
