% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerFitFA.R, R/hello.R
\name{rpg}
\alias{rpg}
\title{Person fit index: Personal Correlation}
\usage{
rpg(X)

rpg(X)
}
\arguments{
\item{X}{Matrix with responses related to n participants to m items (n x m)}
}
\value{
Vector with personal correlation for each participant      (n x 1)
}
\description{
`rpg`returns the personal correlation as proposed by Fowler (1954) and
Donlon & Fisher (1968). It is computed as the correlation between the
respondent’s response vector and the vector of item sample means.

It is implemented as it appears in Ferrando PJ, Vigil-Colet A and
Lorenzo-Seva U (2016) Practical Person-Fit Assessment with the Linear FA
Model: New Developments and a Comparative Study. Front. Psychol. 7:1973.
doi: 10.3389/fpsyg.2016.01973
}
\examples{
n <- 3 # persons (3)
m <- 5 # items (5)
mu <- 0 # mean
sigma <- 1 # sd

df <- matrix(data = NA, nrow = n, ncol = m) # raw data
for (i in 1:nrow(df)) {
  df[i,] <- rnorm(n = m, mean = mu, sd = sigma) #m because I organize it by rows
}

rpg(df)
}
\author{
David Gallardo-Pujol
}
